+++ ui/src/features/ui/index.ts
@@ -8,7 +8,9 @@ export * from './field'
 export * from './form'
 export * from './group-select'
 export * from './heading'
+export * from './icon-button'
 export * from './input-or-select'
+export * from './json-code'
 export * from './layout'
 export * from './loader'
 export * from './modal'
